+++ tests/Loader/Dependents/GraphTest.php
@@ -25,6 +25,7 @@ class GraphTest extends TestCase
                 Name::of('vendor/root'),
                 Version::of('1.0.0'),
                 Url::of('http://example.com'),
+                Url::of('http://example.com'),
                 Set::of(new Relation(
                     Name::of('rand/om'),
                     new Constraint('~1.0'),
@@ -34,6 +35,7 @@ class GraphTest extends TestCase
                 new Package(
                     Name::of('vendor/libA'),
                     Version::of('1.0.0'),
+                    Url::of('http://example.com'),
                     Url::of('http://example.com'),
                     Set::of(
                         new Relation(
@@ -50,6 +52,7 @@ class GraphTest extends TestCase
                     Name::of('vendor/libB'),
                     Version::of('1.0.0'),
                     Url::of('http://example.com'),
+                    Url::of('http://example.com'),
                     Set::of(
                         new Relation(
                             Name::of('vendor/root'),
@@ -64,6 +67,7 @@ class GraphTest extends TestCase
                 new Package(
                     Name::of('watev/foo'),
                     Version::of('1.0.0'),
+                    Url::of('http://example.com'),
                     Url::of('http://example.com'),
                     Set::of(
                         new Relation(
@@ -83,6 +87,7 @@ class GraphTest extends TestCase
                 new Package(
                     Name::of('vendor/libC'),
                     Version::of('1.0.0'),
+                    Url::of('http://example.com'),
                     Url::of('http://example.com'),
                     Set::of(),
                 ),